Cleaned up some files
